resolve compilation issues on linux (doesn't link yet)
