+++ rhythmium/src/cef_app.rs
@@ -67,10 +67,12 @@ pub fn get_settings() -> Settings {
         root_cache_path: Some(
             "/Users/mason/Source/github.com/maboesanman/rhythmium/cache_root".to_string(),
         ),
+        #[cfg(target_os = "macos")]
         framework_dir_path: Some(
             "/Users/mason/Source/github.com/maboesanman/rhythmium/build/lib/Frameworks/Chromium Embedded Framework.framework"
                 .to_string(),
         ),
+        #[cfg(target_os = "macos")]
         main_bundle_path: Some(
             "/Users/mason/Source/github.com/maboesanman/rhythmium/build/lib/rhythmium_partial_bundle.app"
                 .to_string(),
