+++ rust_cef/src/structs/window_info.rs
@@ -23,11 +23,18 @@ impl From<&WindowInfo> for cef_window_info_t {
         cef_window_info_t {
             window_name: cef_string::str_into_cef_string_utf16(&val.window_name),
             bounds: val.bounds.into(),
-            hidden: wrap_boolean(val.hidden),
-            parent_view: ptr::null_mut(),
+            #[cfg(target_os = "linux")]
+            parent_window: 0,
             windowless_rendering_enabled: wrap_boolean(val.windowless_rendering_enabled),
             shared_texture_enabled: wrap_boolean(false),
             external_begin_frame_enabled: wrap_boolean(val.external_begin_frame_enabled),
+            #[cfg(target_os = "linux")]
+            window: 0,
+            #[cfg(target_os = "macos")]
+            hidden: wrap_boolean(val.hidden),
+            #[cfg(target_os = "macos")]
+            parent_view: ptr::null_mut(),
+            #[cfg(target_os = "macos")]
             view: ptr::null_mut(),
         }
     }
